update PQ dims to u32

+++ crates/base/src/global/vecf16_l2.rs
@@ -91,8 +91,8 @@ impl GlobalProductQuantization for Vecf16L2 {
         "aarch64+neon"
     ))]
     fn product_quantization_distance<'a>(
-        dims: u16,
-        ratio: u16,
+        dims: u32,
+        ratio: u32,
         centroids: &[F16],
         lhs: Vecf16Borrowed<'a>,
         rhs: &[u8],
@@ -117,8 +117,8 @@ impl GlobalProductQuantization for Vecf16L2 {
         "aarch64+neon"
     ))]
     fn product_quantization_distance2(
-        dims: u16,
-        ratio: u16,
+        dims: u32,
+        ratio: u32,
         centroids: &[F16],
         lhs: &[u8],
         rhs: &[u8],
@@ -143,8 +143,8 @@ impl GlobalProductQuantization for Vecf16L2 {
         "aarch64+neon"
     ))]
     fn product_quantization_distance_with_delta<'a>(
-        dims: u16,
-        ratio: u16,
+        dims: u32,
+        ratio: u32,
         centroids: &[F16],
         lhs: Vecf16Borrowed<'a>,
         rhs: &[u8],
